Android: More accurate Android Sdk location error messages
If the user installs a fresh Android SDK and selects its path in
the Android options, Qt Creator will still say that this is not
a valid Android SDK.

The reason is that Qt Creator also checks for the platform tools
to be installed. Installing those is a separate step which needs
to be done after installing the SDK.

This patch enables Qt Creator to tell the user if the platform
tools are missing, but the SDK is otherwise fine.

Change-Id: I3557fb93d46e8677498843250302d12c8babb1df
Reviewed-by: Daniel Teske <daniel.teske@digia.com>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@digia.com>