ProjectExplorer: Do not crash when the user deletes a RC
Do not crash when the user deletes the RC that was active when
clicking the run button before Creator gets to start it.

This can when somebody starts to clean up his configuration while
being bored waiting for the build/deployment to finish.

This is the critical part of QTCREATORBUG-11594. A smaller issue
in that bug is that the custom RC should not get removed in the
first place.

Task-number: QTCREATORBUG-11594
Change-Id: I9fca542994db8877dffc4d92872f94fdf6bda3d1
Reviewed-by: Daniel Teske <daniel.teske@digia.com>