BlackBerry: Do not add bar-descriptor.xml to .pro file
The bar-descriptor.xml file will no longer be added to the .pro
file as OTHER_FILES, as it was not necessarily the file listed in
the .pro file that was used for creating the package later on,
as that is configured in the deployment settings.

Instead, the file configured in the deployment settings is now
added to the file project tree, and kept in sync if the user
changes the file to use for deployment.

Task-number: QTCREATORBUG-10054
Change-Id: I3068fedf8ea17f4e60131f9106b3844086d1552b
Reviewed-by: David Kaspar <dkaspar@blackberry.com>
Reviewed-by: Mehdi Fekari <mfekari@blackberry.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>