Vcs: Prettify output in CheckoutProgressWizardPage
Handle CR in the output of the VCS we are running in the checkout
wizards. This makes sure we get proper progress information when
running the checkout operation.

Note that this is not a perfect implementation: It will fail when
e.g. only a number at the start of the line is updated and the
rest of the text in the line is reused.

Task-number: QTCREATORBUG-10112
Change-Id: If742e5cb945a2fcada8319d08610d1ccc7fa2ae8
Reviewed-by: Christian Stenger <christian.stenger@digia.com>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Robert Loehning <robert.loehning@digia.com>
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>