Qnx: Use correct mkspec in the qmake step
In order to use the correct mkspec in the qmake step,
the qt version mkspec needs to be explicitly listed in the QnxToolChain
suggested mkspec list (Otherwise the first one found in
list will be used).

That is currently broken due to the following issues:

* The mkspec for BB armle-v7 target has been recently
changed to blackberry-armle-v7-qcc, which was not listed
in the QnxToolChain suggested mkspec list.

* The Qt4 for BB armle-v7 uses the "default" mkspec which is a copy
of the "blackberry-armv7le-qcc" mkspec. The mkspec
name should be explicit in order to be found and used.

Task-number: QTCREATORBUG-11674

Change-Id: Idb53d22e7c4b746b23dd3a65c42d00dca5c27d0c
Reviewed-by: Daniel Teske <daniel.teske@digia.com>
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>