AndroidManifestEditor: Rewrite writing of xml files
Using QXmlStreamWriter + QXmlStreamReader gives us greater control
over the ordering of attributes. That way we can ensure that we never
change the ordering that the user used. (Whereas for QDom the best
we can do is to have a predictable but random order.)

Neither QXmlStreamWriter nor QDom allows us to control the indentation
and other whitespace, so any user indentation is lost.

In addtion QXmlStreamWriter seems to have several bugs with indentation
though, leading to rather strangely formatted files.

Task-number: QTCREATORBUG-10870

Change-Id: I4cdbdcd499227f418e7767eb1b532efbbd5083a5
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>
Reviewed-by: Alessandro Portale <alessandro.portale@digia.com>