ProjectTree; Fix directoryFor() to handle new resource nodes
The resource nodes do have a special path, the code assumed that
folder's path were always directories. This is no longer true.

Both pathFor() and directoryFor() suffer from path() being special for
VirtualFolderNodes and ResourceNodes. That specialness is probably not
a good idea, cleaning that up is a longer term task.

Task-number: QTCREATORBUG-11677
Change-Id: I995e82b7ed361b5e2c6681484074c65e2175c96e
Reviewed-by: Daniel Teske <daniel.teske@digia.com>