ios: simulator support for Xcode 5.1
update iossim to be based on version 1.9.0, so that we can support
the new Xcode 5.1. Keep the old version around and use it
with older Xcode.
All the logic has been moved into iossim which does an exec of the
older binary if needed (so that creator itself remains untouched)

Task-number: QTCREATORBUG-11714
Change-Id: I4ad9f922f22367c8371b3be68eae5b6dfdf4df5c
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@digia.com>