Doc: Use the page title for linking to index.html
Because the Qt Creator documentation depends on Qt reference
docs for linking, qdoc will generate a collision page for
index.html as it exists in both.

This change prevents linking to the collision page in the
navigation links by using the page title as the link target
instead of the file name.

Task-number: QTCREATORBUG-11035
Change-Id: I5638e09b60543dc2068b0e4a269f8f72d4655b48
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@digia.com>