Update debugger language aspect in run settings when qmake step changes
When QML language debugging hasn't been set to a fixed state yet by the
user try to find a qmake step to decide whether to enable QML debugging by default
or not.

This is a hack, breaking the separation between qt build steps and debugger
run settings. However, adding a generic project infrastructure for this specific
use case is probably overkill...

Task-number: QTCREATORBUG-11474
Change-Id: Ib65c8474b9b7ec187769c209531ff56bc8293cde
Reviewed-by: Aurindam Jana <aurindam.jana@digia.com>
Reviewed-by: Eike Ziller <eike.ziller@digia.com>
Reviewed-by: Alessandro Portale <alessandro.portale@digia.com>