WelcomePage: Support additional (non-qt) example sets
Added with setting "Help/InstalledExamples" which is a string list, each
string is a triple of (display name, manifest parse path, examples
path), joined with '|'.
If a documentationPath for a Qt version is already added through the
InstalledExamples setting, we do not show an extra combo box item for
it. That makes it possible for an installer to register a single example
set for the same Qt version in different variants (e.g. iOS, Android
x86, Android arm)

Change-Id: I42dd8f9ca9cb6cee247e3dc8ce9d014e56852e79
Reviewed-by: Kai Koehne <kai.koehne@digia.com>