QtVersionManager: Only load Qt versions after the toolchains are available
The mingw Qt versions need to know the mingw toolchains to run qmake since
that may link to a library provided by mingw.

This patch delays loading of the Qt versions till after the Tool Chains are
available by explicitly listening to the toolChainsLoaded signal of the
Toolchain Manager.

I do not see how else we can enforce the proper ordering with the plugin
initialization scheme we have.

Task-number: QTCREATORBUG-11898
Change-Id: I5a93c2b2b32c658695017295652242a5aaa6ee60
Reviewed-by: Daniel Teske <daniel.teske@digia.com>
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>