Qnx: Properly setup the environment vars in the bar descriptor
The environment variables in the bar descriptor are modified
depending on the Qt deployment option the user selects.

This causes inconsistent behavior if the user modifies those values
on the template file. The env var values will be generated
in the final bar descriptor without being visible on the template
to avoid any user modifications that won't apply correclty.

Note: For next releases, the bar descriptor needs to be handled
properly to avoid any black magic so all modifications should
be visible to the user.

Task-number: QTCREATORBUG-11916
Change-Id: I00a2c855be3def31cd008071c6269e5a69f81536
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>