QmlDesigner: Always warn asynchronously in exceptions
Exceptions are likely to be triggered from QML code.
Calling QMessageBox::warning() results in a call to exec() and
exec() returns back into the Qt event loop.

This is dangerous, because QML might have scheduled a deleteLater
which is then executed by exec(). This can crash the application, because
the object is supposed to be "deleted later" and not directly
as a side effect of the signal that was emmited by QML.

Task-number: QTCREATORBUG-11946

Change-Id: Icb884847633bd421d3dc7bc034d598ef84d39429
Reviewed-by: Marco Bubke <marco.bubke@digia.com>