support the latest g++ on mac
the latest g++ on mac is really clang with a compatibility frontend,
and claims to be clang and support all the features clang does
with -std=c++0x (which qmake by default passes in).
But if you link libstdc++ (again the default) then the library has
several gaps. Catch that.

Change-Id: I8387f8fdcfc7639538e576a2e93b2301e07086cd
Reviewed-by: Eike Ziller <eike.ziller@digia.com>
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@digia.com>
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@digia.com>