Open projects from file system view.
Use ICore::openFiles() instead of
EditorManager::openEditor() in FolderNavigationWidget
since it is also able to open projects.

Add menu entry to open a project in a directory, which
uses a logic similar to the command line processing,
trying to locate a project file in the directory
using file globs.

This allows for conveniently opening several projects by
right-clicking instead of having to open the file dialog
and to look for .pro files in directory listings.

Change-Id: I8404342f7f09b4b994e6eee82af3dc712cd97512
Reviewed-by: Daniel Teske <daniel.teske@digia.com>