When the qbs submodule is present, build its applications too.
So far, we have only built the library, arguing that that's all
QtC needs from qbs. However, it turns out that it is useful to have
the qbs executable as well, e.g. if there is a problem and the
user wants to quickly compare the behavior of QtC and the
command line tool with the guarantee that they're using the exact
same qbs library.

Task-number: QBS-558
Change-Id: Iddf94a0b304467abee313bcd52e446f02156c98b
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>