Valgrind: Fix display name for analyzing remote external application.
The current one just adds the string "(Remote)", which is pretty much
guaranteed to trick users into thinking it's the one they must choose
when running remote valgrind. Instead, we must make it clear that it's
for analyzing something other than the current project. Unfortunately,
the string still contains the word "remote" because for some reason,
we don't offer valgrind for local external applications.

Change-Id: I27296851b747437e96dc22dbaf54eaf10dcf4be5
Reviewed-by: hjk <hjk121@nokiamail.com>