QmlDesigner.PropertyEditor: Fix the small color button for gradients
The small ColorCheckButton has to preview the same color as the rest
of the color editor. Before this patch it was showing the color
defined in the backend as rectangle color.
This is not correct in gradient editing mode.
Also the binding in Colorbutton was deleted assigning values
directly.

Task-number: QTCREATORBUG-12204
Change-Id: Iad5263ccccba10fdb4f2b757f6965d500c50ac53
Reviewed-by: Christian Stenger <christian.stenger@digia.com>