BaseFileWizard: Get rid of applyExtensionPageShortTitle(...)
That method was used to set a title in the progress view of the wizard
that is different from the page title. That is used exactly once and
there it adds more confusion than it helps. So get rid of the whole
thing.

Consistently set the "shortTitle" property instead for all wizards
that want to have a separate short title.

Change-Id: Ia4881e9c00891058629491f9e9de4ac421c59727
Reviewed-by: Orgad Shaneh <orgads@gmail.com>