Rename ICustomWizardFactory to ICustomWizardMetaFactory
Also rename CustomWizardFactory to CustomWizardMetaFactory. These factories
are used to generate the actual factories that are visible in File->New, so
they are kind of Meta:-)

CustomWizardClassFactory was my first attempt, but since there are project-
and classwizards that are generated by this meta-factory that is not a good
name.

Get rid of a useless typedef while at it.

Change-Id: I3dc2cce3f52999af7e59d5f2ad5509153056ea49
Reviewed-by: Orgad Shaneh <orgads@gmail.com>