Qnx: Generate bar descriptor when opening new projects
When opening a project the first time with Qt Creator
and selecting a BlackBerry kit, Qt Creator does not pop up
the dialog asking the user to add a bar descriptor. This
makes the deployment step useless since the bar descriptor
is mandatory for packaging.

The user has to close and reopen the project again in order
to be able to generate a bar-descriptor.

Change-Id: Ibb930e60d3b9d98ad9631caa092f6ec17b33b749
Reviewed-by: Sergio Ahumada <sahumada@blackberry.com>
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>