QmlDesigner.PropertyEditor: Fix TextEditSpecifics
I moved the property format into Text section, because
having two sections with even the same name was confusing.
The format property also is not really related to "Text Input".

Since the distinction between plain and rich text is very important,
it makes sense to have this property in the first section.

Change-Id: I4d66ed4a24fe0d99d394c17d149c830db2c6d3b8
Task-number: QTCREATORBUG-12452
Reviewed-by: Tim Jenssen <tim.jenssen@digia.com>