QmlDesigner: Ensure we always have a document for a .qml file.
In some cases there is no document in the snapshot and
Qt Quick Designer just ignored the new document in this case.

So for example when a new tab was added and then opened the first time,
there was no document in the snapshot and Qt Creator opened the document
without Qt Quick Designer reacting to this.

Change-Id: I98f97c2443686733b7e71d852c6df63865034dfc
Reviewed-by: Marco Bubke <marco.bubke@digia.com>