QmlDesigner.PropertyEditor: Make anchoring consistent
If we anchor left or right we have to disable horizontal
anchoring.
Actually this is still simpler then what Qt Quick actually
supports, since Qt Quick does support anchoring to the left
and horizontal anchoring.
What is not supported is anchoring to the left, right and
horizontally.
This has to be fixed.

Change-Id: Id38e528d9d8a97961135370b0c9c0151bd4b6f18
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@digia.com>