Change the scene graph profiler to give more accurate information
The scene graph profiler was pretty barren. The ranges it showed were
only vaguely related to actual events and the numbers shown in the
labels weren't really informative.

This change shows every stage in GUI thread and rendering thread as its
own range, so that you can easily see now if a particular stage takes
unusually long. This comes at a cost, though: The color is used for
showing event types instead of frame rates now and there is no 1:1
mapping of rows to threads anymore. The first row is still reserved
for the GUI thread and the second one for the render thread. Further
rows are used as needed, though, and you can meaningfully expand the
category now.

Task-number: QTBUG-39876
Change-Id: Ib91ba454dd0e3e7482ca2afe6d3c80d9868bcce3
Reviewed-by: Kai Koehne <kai.koehne@digia.com>