Fix that editor actions were enabled when help window is active
The editor manager actions should not be enabled just because the main
window is in edit/debug/analyze mode. Instead the corresponding mode
must have focus.
Otherwise they will be enabled even if extra windows that do not have
editors (like the extra help window, which is e.g. used for
the example documentation) are active.

Change-Id: Id1bf3cc4d1a761ef7ea8a275701f78626ff44533
Reviewed-by: David Schulz <david.schulz@digia.com>