QmlProfiler: Rename Qt5 "Painting" to "Animations" and always show it
"Animations" is a much better name as that's what it actually shows. It
also serves to distinguish from the QtQuick 1 only "Painting" category
that really does show the time QPainter took to paint the scene.

Hiding the "Animations" category when profiling QtQuick 1 makes no
sense as we also don't hide the JavaScript category or any of the
commercial addons which are only available from QtQuick 2. As QtQuick 1
is deprecated we keep hiding the Painting category when profiling
QtQuick 2 to reduce clutter in the UI for the more common case.

Change-Id: I0e14fab8abf427dc1abb6d00874e7e4af0fa30e5
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@digia.com>
Reviewed-by: Kai Koehne <kai.koehne@digia.com>