ManhattanStyle: StyleHelper::drawArrow() dpi aware
This patch makes the drawing and caching of the
QStyle::PE_IndicatorArrow* elements devicePixelRatio-aware. Also, the
triangle and shadow color are adjusted to be consistent with the other
toolbar icons.

The custom triangle drawig code (which was copied from a base style) has
been replaced by calls of QCommonStyle::drawPrimitive(). A filled
45-45-90 triangle is best rendered without antialiasing.

Change-Id: I9a4371ec4dc04bd4455050913219c50adfcb37b5
Reviewed-by: Alessandro Portale <alessandro.portale@digia.com>