Core: Use enable the use of magnifier@2x.png
magnifier@2x.png was already part of the core plugin, but never used
since the @2x logic is in QIcon and for the magnifier.png we use QPixmap
for loading.
This patch uses Utils::StyleHelper::dpiSpecificPixmap() to load the
image in the findtoolbar and locatorwidget. It also removes the image
centering code in FindToolBar::updateIcons() (most likely a by now
obsolete workaround) since the image gets properly centered in the
Utils::FancyLineEdit.

Change-Id: I5cd6d22fc5cce3838db6164568cb9165da052c49
Reviewed-by: Eike Ziller <eike.ziller@digia.com>