New Dialog: Do not show wizards on top of everything else
Just live with the fact that there is not even a way to keep it on top
of the main window or something like that (without also making it a tool
window).
Also, even though it would be nice to have the "New" menu item bring the
current wizard to the top, if any, that requires heavy refactoring in
the wizard side.

Change-Id: I93d345964b085adeaa314d93acc612f8fb2e7e68
Task-number: QTCREATORBUG-12675
Reviewed-by: Daniel Teske <daniel.teske@digia.com>