qmldesigner: fix reparenting to Layout
Using strings for everything creates issues, the whole thing should
be cleaned up in master.
cleverCompare is asymmetric, one side has the code model, but the other
is a string.
To keep it like that wherever we detect that the string is becoming
too strange (normally for copied prototypes) we fall back to the
cpp name if available whenever the import does not match the origin.
Then cleverCompare can successfully compare the strings.

Task-number:QTCREATORBUG-11905

Change-Id: I857f81b57d8ddd43263ce0e3bb085600ff6e8a43
Reviewed-by: Tim Jenssen <tim.jenssen@digia.com>