Improve AndroidManifest.xml error dialog content
Currently, if the AndroidManifest.xml file is not present in the
installation folder (e.g. failed make install), the error message
shown to the user just tells him that some operation failed. This patch
tries to be more verbose to give the user some clues for the failure
reason.

Task-number: QTCREATORBUG-11503
Change-Id: I013de394c87b3adb53ec86dd97433567d7f63049
Reviewed-by: Daniel Teske <daniel.teske@digia.com>