Doc: update info about setting Qt in PATH
Use a variable for the path to the bin directories,
because the example PATH is most likely wrong.
Add a link to "Using Command Line Options" and move the
note there, because it is advanced usage (needed only for
self-built Qt Creator). However, these developers will
probably know how to set the PATH.

Change-Id: I6d6563f0c7d6aebcc4a5887f9cbb86b5d7e1676a
Reviewed-by: Eike Ziller <eike.ziller@digia.com>
Reviewed-by: hjk <hjk121@nokiamail.com>