ImageViewer: No interpolation when zooming in
The imageViewer currently always renders the images smooth/interpolated.
When loading an image in an IDE and zooming in, I usually do that in
order to examine the details on pixel level. Therefore, I believe that
smoothing/interpolation should be turned off with a scalefactor >= 1.

This patch does that for images and movies. Svg files remain always
smooth.

Change-Id: I327583b1f46132d179a1e0425f146106d962377b
Reviewed-by: Daniel Teske <daniel.teske@digia.com>