FeatureProvider: Move features into Kits
Make available features and platforms methods on kits (implemented
via the KitInformation). Add convenience methods to KitManager to
collect all that.

Remove QtFeatureProvider and implement KitFeatureProvider instead.

The idea of this patch is to make it easier to find which kits are
applicable to which wizard: I should now be able to match the
wizard features against a kit and have a good set of kits to choose
from.

Change-Id: Ie5be0213f142cfdf4417ac55bd6cbb056265a531
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@digia.com>